fix: edge function 수정 및 todo_list 생성 로직 수정
